+++ config/kit.php
@@ -15,7 +15,7 @@ return [
     | contra a árvore de trabalho, que é mais ruidosa.
     */
 
-    'version' => '0.9.3',
+    'version' => '0.9.4',
 
     /*
     |--------------------------------------------------------------------------